feat: added input for all parameters except path

+++ src/main/java/backend/academy/fractal/flame/model/records/FractalConfig.java
@@ -10,7 +10,7 @@ public record FractalConfig(
     Optional<List<Color>> colors,
     int degreeOfRandomnessOfFractalCreation,
     boolean complicateFlameShape,
-    short iterPerSample,
+    int iterPerSample,
     int samples
 ) {
 }
